refactor: 增加 BootstrapInputNumberBase 基类
