+++ src/BootstrapBlazor/Components/InputNumber/BootstrapInputNumber.razor.cs
@@ -91,24 +91,6 @@ public partial class BootstrapInputNumber<TValue>
     private IIconTheme? IconTheme { get; set; }
 
     /// <summary>
-    /// SetParametersAsync 方法
-    /// </summary>
-    /// <param name="parameters"></param>
-    /// <returns></returns>
-    public override Task SetParametersAsync(ParameterView parameters)
-    {
-        // Unwrap Nullable<T>, because InputBase already deals with the Nullable aspect
-        // of it for us. We will only get asked to parse the T for nonempty inputs.
-        var targetType = Nullable.GetUnderlyingType(typeof(TValue)) ?? typeof(TValue);
-        if (!typeof(TValue).IsNumber())
-        {
-            throw new InvalidOperationException($"The type '{targetType}' is not a supported numeric type.");
-        }
-
-        return base.SetParametersAsync(parameters);
-    }
-
-    /// <summary>
     /// <inheritdoc/>
     /// </summary>
     protected override void OnParametersSet()
