+++ src/BootstrapBlazor/Components/InputNumber/BootstrapInputNumberBase.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) Argo Zhang (argo@163.com). All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+// Website: https://www.blazor.zone or https://argozhang.github.io/
+
+namespace BootstrapBlazor.Components;
+
+/// <summary>
+/// BootstrapInputNumber 基类
+/// </summary>
+public class BootstrapInputNumberBase<TValue> : BootstrapInputBase<TValue>
+{
+    /// <summary>
+    /// SetParametersAsync 方法
+    /// </summary>
+    /// <param name="parameters"></param>
+    /// <returns></returns>
+    public override Task SetParametersAsync(ParameterView parameters)
+    {
+        // Unwrap Nullable<T>, because InputBase already deals with the Nullable aspect
+        // of it for us. We will only get asked to parse the T for nonempty inputs.
+        var targetType = Nullable.GetUnderlyingType(typeof(TValue)) ?? typeof(TValue);
+        if (!typeof(TValue).IsNumber())
+        {
+            throw new InvalidOperationException($"The type '{targetType}' is not a supported numeric type.");
+        }
+
+        return base.SetParametersAsync(parameters);
+    }
+}
